langchain-agent/requirements.txt
langchain
langchain-google-genai
streamlit
python-dotenv
requests
PyYAML
